Added Activity_01_D updated variable data types

+++ src/main/kotlin/activity_01_d/Main.kt
@@ -1,33 +1,32 @@
 package activity_01_d
 
 fun main(){
-    var amount: Double = 0.0
-    var divisor: Double = 0.0
+    var amount: Int = 0
+    var divisor: Int = 0
 
     fun getAmount(){
-        var input: Double? = 0.0
+        var input: Int? = 0
         println("Please Enter 5 Numbers")
         try {
             for(i in 1..5){
-                input = readLine()!!.toDouble()
+                input = readLine()!!.toInt()
                 amount += input
             }
 
         }catch (exception: Exception){
             println("Invalid Input!! Try Again")
-            amount = 0.0
+            amount = 0
             getAmount()
         }
     }
-
-    fun calculateAmount(divisor: Double){
+    fun calculateAmount(divisor: Int){
         amount /= divisor
         println(amount)
     }
 
     getAmount()
     println("Divide the value by how many?")
-    divisor = readLine()!!.toDouble()
+    divisor = readLine()!!.toInt()
     calculateAmount(divisor)
 
 }
